+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## [v2.0.1](https://github.com/ganrad/openai-api-router/compare/v2.0.0...v2.0.1)
+### Functionality changes
+* (Bugfix) Resolved JSON parse issue for streamed chat completion response.
+* (Enhancement) Simplified logic for streamed responses.  The modified logic is more performant and enhances maintainability.
+* (Bugfix) Auth strategy will only be configured when authenication is enabled for the gateway.  This was causing a server startup issue.
+
 ## [v2.0.0](https://github.com/ganrad/openai-api-router/compare/v1.9.0...v2.0.0)
 ### Functionality changes
 * (Enhancement) The REST API's exposed by AI Application Gateway solution can now be secured using Microsoft Entra ID.  This feature is enabled by default.